requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
argcomplete==3.6.3
certifi==2026.2.25
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
commitizen==4.13.9
decli==0.6.3
Deprecated==1.3.1
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.1
filelock==3.25.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.17
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
nodeenv==1.10.0
openai==2.26.0
packaging==26.0
platformdirs==4.9.4
pre_commit==4.5.1
prompt_toolkit==3.0.51
pydantic==2.12.5
pydantic_core==2.41.5
python-discovery==1.1.0
PyYAML==6.0.3
questionary==2.1.1
ruff==0.15.5
sniffio==1.3.1
starlette==0.52.1
termcolor==3.3.0
tomlkit==0.14.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
virtualenv==21.1.0
wcwidth==0.6.0
wrapt==2.1.1
